+++ CHANGELOG.md
@@ -12,6 +12,7 @@ We use _breaking :warning:_ to mark changes that are not backward compatible (re
 ## Unreleased
 
 ### Added
+- [#7](https://github.com/prmsrswt/foundry/pull/7) Furnace: Add gRPC protobuf. Add `foundry furnace` sub-command to run the Furnace component.
 
 ### Fixed